refactor: Deleted unnecessary DTOs and moved the existing ones to the model package refactor: Change GithubAPI from a class to an object
