--- app/src/main/java/pl/lambada/songsync/data/dto/GithubReleaseDTO.kt
@@ -1,13 +1,0 @@
-package pl.lambada.songsync.data.dto
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class GithubReleaseResponse(
-    @SerialName("html_url")
-    val htmlURL: String,
-    @SerialName("tag_name")
-    val tagName: String,
-    val body: String
-)
