+++ app/src/main/java/pl/lambada/songsync/domain/model/lyrics_providers/others/LRCLib.kt
@@ -1,4 +1,4 @@
-package pl.lambada.songsync.data.dto
+package pl.lambada.songsync.domain.model.lyrics_providers.others
 
 import kotlinx.serialization.Serializable
